requirements.txt
altair==4.2.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
charset-normalizer==3.0.1
click==8.1.3
decorator==5.1.1
entrypoints==0.4
ffmpeg-python==0.2.0
filelock==3.9.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
more-itertools==9.0.0
numpy==1.24.2
openai-whisper==20230124
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytube==12.1.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.18.1
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
tornado==6.2
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
zipp==3.14.0
